requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
dataclasses-json==0.5.1
discord.py==1.3.3
idna==2.9
marshmallow==3.6.1
marshmallow-enum==1.5.1
multidict==4.7.5
mypy-extensions==0.4.3
pydle==0.9.4
stringcase==1.2.0
typing-extensions==3.7.4.2
typing-inspect==0.6.0
websockets==8.1
yarl==1.4.2
